requirements.txt
matplotlib==3.7.1
matplotlib-inline==0.1.6
plotly==5.15.0
pytorch-lightning==2.0.4
rich==13.4.2
scikit-learn==1.2.0
click==8.1.3
imageio==2.31.1
numpy==1.22.2
optuna==3.3.0
plotly==5.15.0
PyYAML==6.0
kaleido==0.2.1
statsmodels==0.14.2
PyYAML==6.0